modif update

+++ Project1/Node.hpp
@@ -5,11 +5,11 @@
 
 struct Node {
     sf::Vector2i position;
-    float gCost, hCost, fCost;
+    int gCost, hCost, fCost;
     Node* parent;
 
     Node(sf::Vector2i pos);
-    void calculateCosts(Node* end, float newG);
+    void calculateCosts(Node* end, int newG);
 };
 
 #endif
